third_party/blink/web_tests/external/wpt/css/css-fonts/parsing/font-face-src-tech-expected.txt

This is a testharness.js-based test.
PASS Check that src: url("foo.ttf") is valid
PASS Check that src: url("foo.ttf") tech() is invalid
FAIL Check that src: url("foo.ttf") tech(features-opentype) is valid assert_equals: expected true but got false
FAIL Check that src: url("foo.ttf") tech(features-aat) is valid assert_equals: expected true but got false
FAIL Check that src: url("foo.ttf") tech(color-COLRv0) is valid assert_equals: expected true but got false
FAIL Check that src: url("foo.ttf") tech(color-COLRv1) is valid assert_equals: expected true but got false
FAIL Check that src: url("foo.ttf") tech(color-sbix) is valid assert_equals: expected true but got false
FAIL Check that src: url("foo.ttf") tech(color-CBDT) is valid assert_equals: expected true but got false
FAIL Check that src: url("foo.ttf") tech(variations) is valid assert_equals: expected true but got false
FAIL Check that src: url("foo.ttf") tech(palettes) is valid assert_equals: expected true but got false
PASS Check that src: url("foo.ttf") tech("features-opentype") is invalid
PASS Check that src: url("foo.ttf") tech("color-COLRv0") is invalid
PASS Check that src: url("foo.ttf") tech("variations") is invalid
FAIL Check that src: url("foo.ttf") tech(features-opentype, color-COLRv0, variations, palettes) is valid assert_equals: expected true but got false
PASS Check that src: url("foo.ttf") tech(features-opentype color-COLRv0 variations palettes) is invalid
PASS Check that src: url("foo.ttf") tech(auto) is invalid
PASS Check that src: url("foo.ttf") tech(default) is invalid
PASS Check that src: url("foo.ttf") tech(inherit) is invalid
PASS Check that src: url("foo.ttf") tech(initial) is invalid
PASS Check that src: url("foo.ttf") tech(none) is invalid
PASS Check that src: url("foo.ttf") tech(normal) is invalid
PASS Check that src: url("foo.ttf") tech(xyzzy) is invalid
FAIL Check that src: url("foo.ttf") format(opentype) tech(features-opentype) is valid assert_equals: expected true but got false
PASS Check that src: url("foo.ttf") tech(features-opentype) format(opentype) is invalid
FAIL Check that src: url("foo.ttf") tech(incremental), url("bar.html") is valid assert_equals: expected true but got false
FAIL Check that src: url("foo.ttf") tech(incremental, color-SVG, features-graphite, features-aat), url("bar.html") is valid assert_equals: expected true but got false
FAIL Check that src: url("foo.ttf") tech(color-SVG, features-graphite), url("bar.html") is valid assert_equals: expected true but got false
FAIL Check that src: url("foo.ttf") tech(color-SVG), url("bar.html") is valid assert_equals: expected true but got false
FAIL Check that src: url("foo.ttf") tech(features-graphite), url("bar.html") is valid assert_equals: expected true but got false
PASS Check that src: url("foo.ttf") dummy("opentype") tech(variations) is invalid
PASS Check that src: url("foo.ttf") dummy("opentype") dummy(variations) is invalid
PASS Check that src: url("foo.ttf") format(opentype) tech(features-opentype) dummy(something) is invalid
PASS Check that src: url("foo.ttf") format(dummy), url("foo.ttf") tech(variations) is invalid
PASS Check that src: url("foo.ttf") tech(color, url("bar.html") is invalid
Harness: the test ran to completion.
